+++ examples/nodejs/env.js
@@ -1,7 +1,0 @@
-const ACCOUNT_ADDRESS = '';
-const ACCOUNT_PRIVATEKEY = '';
-
-module.exports = {
-  ACCOUNT_ADDRESS,
-  ACCOUNT_PRIVATEKEY,
-};