+++ packages/file-handle/src/go-wasm/main.go
@@ -6,6 +6,7 @@ import (
 	"syscall/js"
 
 	lib "github.com/bnb-chain/greenfield-common/go/hash"
+	redundancy "github.com/bnb-chain/greenfield-common/go/redundancy"
 )
 
 func hashFunc(this js.Value, args []js.Value) interface{} {
@@ -36,7 +37,28 @@ func hashFunc(this js.Value, args []js.Value) interface{} {
 	}
 }
 
+func encodeRawSegment(this js.Value, args []js.Value) interface{} {
+	array := args[0]
+	dataBlocks := args[1].Int()
+	parityBlocks := args[2].Int()
+
+	byteLength := array.Get("byteLength").Int()
+	data := make([]byte, byteLength)
+	var buffer []uint8 = make([]uint8, byteLength)
+	js.CopyBytesToGo(buffer, array)
+	reader := bytes.NewReader(buffer)
+	reader.Read(data)
+
+	encodeShards, _ := redundancy.EncodeRawSegment(data, dataBlocks, parityBlocks)
+	shardsJson, _ := json.Marshal(encodeShards)
+	return map[string]interface{}{
+		"result": string(shardsJson),
+	}
+}
+
 func main() {
+	done := make(chan int, 0)
 	js.Global().Set("getCheckSums", js.FuncOf(hashFunc))
-	<-make(chan struct{})
+	js.Global().Set("encodeRawSegment", js.FuncOf(encodeRawSegment))
+	<-done
 }